+++ src/check/location.rs
@@ -1,0 +1,11 @@
+use serde::Serialize;
+use std::path::PathBuf;
+
+/// the position of an issue or fix
+#[derive(Clone, Debug, Default, Eq, Hash, Ord, PartialOrd, PartialEq, Serialize)]
+pub struct Location {
+    pub file: PathBuf,
+    pub line: u32,
+    pub start: u32,
+    pub end: u32,
+}
