+++ src/check/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) mod issue;
+mod location;
+pub(crate) mod scanners;
+
+pub(crate) use issue::Issue;
+pub(crate) use location::Location;
