+++ src/check/scanners/unordered_sections.rs
@@ -1,4 +1,6 @@
-use crate::{Config, Document, Issue, Location};
+use crate::check::{Issue, Location};
+use crate::database::Document;
+use crate::Config;
 
 /// populates the given issues list with all sections in this document that don't match the configured order
 pub fn scan(doc: &Document, config: &Config, issues: &mut Vec<Issue>) {
@@ -57,7 +59,9 @@ pub fn scan(doc: &Document, config: &Config, issues: &mut Vec<Issue>) {
 #[cfg(test)]
 mod tests {
 
-    use crate::{Config, Document, Issue, Location};
+    use crate::check::{Issue, Location};
+    use crate::database::Document;
+    use crate::Config;
     use indoc::indoc;
     use std::path::PathBuf;
