+++ src/commands/mod.rs
@@ -7,18 +7,18 @@ mod json_schema;
 mod pitstop;
 mod stats;
 
-use crate::fix::Fix;
+// re-exports
 pub use check::check;
 pub use fix::fix;
 pub use init::init;
 pub use json_schema::json_schema;
 pub use pitstop::pitstop;
-use serde::Serialize;
 pub use stats::stats;
-use std::path::PathBuf;
 
-// TODO: move to a new "check" module
-// TODO: move the contents of "scan" into a "check/scanners" module
+// used locally
+use crate::check::Issue;
+use crate::Fix;
+
 /// The inner API of the check subsystem.
 /// This data structure is returned by the probes.
 /// It contains highly structured, semantically meaningful data
@@ -39,151 +39,4 @@ impl Outcome {
             fixes: vec![],
         }
     }
-}
-
-/// the issues that this linter can find
-#[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
-pub enum Issue {
-    BrokenImage {
-        location: Location,
-        target: String,
-    },
-    CannotReadConfigurationFile {
-        location: Location,
-        message: String,
-    },
-    CannotWriteConfigFile {
-        message: String,
-        file: PathBuf,
-    },
-    CannotWriteJsonSchemaFile {
-        file: PathBuf,
-        message: String,
-    },
-    DocumentWithoutLinks {
-        location: Location,
-    },
-    DuplicateSection {
-        location: Location,
-        title: String,
-    },
-    EmptyDocument {
-        path: PathBuf,
-    },
-    EmptySection {
-        location: Location,
-        title: String,
-    },
-    InconsistentHeadingLevel {
-        location: Location,
-        /// human-readable section title
-        section_title: String,
-        /// the most commonly observed level (if one exists)
-        common_level: Option<u8>,
-        /// the level used here
-        this_level: u8,
-        /// all observed variants
-        all_levels: Vec<u8>,
-    },
-    InvalidConfigurationFile {
-        location: Location,
-        message: String,
-    },
-    InvalidGlob {
-        glob: String,
-        message: String,
-        location: Location,
-    },
-    InvalidTitleRegex {
-        regex: String,
-        problem: String,
-        file: PathBuf,
-    },
-    LinkToNonExistingAnchorInCurrentDocument {
-        location: Location,
-        /// the non-existing anchor in the current
-        anchor: String,
-    },
-    LinkToNonExistingAnchorInExistingDocument {
-        location: Location,
-        /// the file that the link points to
-        target_file: String,
-        /// the non-existing anchor in that file
-        anchor: String,
-    },
-    LinkToNonExistingFile {
-        location: Location,
-        target: String,
-    },
-    LinkToSameDocument {
-        location: Location,
-    },
-    LinkWithoutTarget {
-        location: Location,
-    },
-    MissingFootnote {
-        location: Location,
-        identifier: String,
-    },
-    MissingLink {
-        location: Location,
-        path: PathBuf,
-        title: String,
-    },
-    MixCapSection {
-        location: Location,
-        all_variants: Vec<String>,
-        this_variant: String,
-        common_variant: Option<String>,
-        section_level: u8,
-    },
-    NoTitleSection {
-        location: Location,
-    },
-    ObsoleteOccurrencesSection {
-        location: Location,
-    },
-    OrphanedResource {
-        // This is a String and not a Path because we need a String (to print it),
-        // and we already converted the Path of this orphaned resource into a String
-        // during processing it.
-        location: Location,
-    },
-    SectionWithoutHeader {
-        location: Location,
-    },
-    TitleRegexNoCaptures {
-        regex: String,
-    },
-    TitleRegexTooManyCaptures {
-        regex: String,
-        captures: usize,
-    },
-    UnclosedBacktick {
-        location: Location,
-    },
-    UnclosedFence {
-        location: Location,
-    },
-    UnknownSection {
-        location: Location,
-        title: String,
-        allowed_titles: Vec<String>,
-    },
-    UnorderedSections {
-        location: Location,
-    },
-    UnusedFootnote {
-        location: Location,
-        identifier: String,
-    },
-}
-
-/// the position of an issue or fix
-#[derive(Clone, Debug, Default, Eq, Hash, Ord, PartialOrd, PartialEq, Serialize)]
-pub struct Location {
-    pub file: PathBuf,
-    pub line: u32,
-    pub start: u32,
-    pub end: u32,
 }
