+++ src/fix/empty_section.rs
@@ -1,7 +1,8 @@
 use super::Fix::RemovedEmptySection;
+use crate::check::Location;
 use crate::fix;
 use crate::fix::Result::Fixed;
-use crate::{Location, Tikibase};
+use crate::Tikibase;
 
 pub fn remove_section(base: &mut Tikibase, title: String, location: Location) -> fix::Result {
     let base_dir = base.root.clone();
