+++ src/fix/unordered_sections.rs
@@ -1,8 +1,9 @@
 use super::Result::Fixed;
+use crate::check::Location;
 use crate::database::Section;
 use crate::fix;
 use crate::fix::Fix::SortedSections;
-use crate::{Location, Tikibase};
+use crate::Tikibase;
 
 pub fn sort_sections(base: &mut Tikibase, location: Location) -> fix::Result {
     let base_dir = base.root.clone();
